improved handling of half-aligned reads
A half-aligned read is one where one mate is aligned and the other
isn't.  We treat the aligned mate exactly as if it was a single read,
especially if we have actual single reads mixed into the input stream.
If there a re no duplicates, we pass the half-aligned pair through, else
we generate an actual single read as consensus.  The logic interacts in
an ugly way with the retention of original reads, so the code is
somewhat intricate and hard to test thoroughly.  So, fingers crossed,
it's there at any rate.