fixed pretzel logic in bam-rmdup
Proximal symptom was that second mates tended to be demoted to unpaired
reads.  The real reason was logic that didn't make sense at all.  This
should be fixed now, though the interaction of the various filtering and
pass-through options is rather counter-intuitive.  Also, Bio.Bam.Rmdup
looks much cleaner now.