fixed generation of MD fields in wrapTo
Originally, I attempted to generated MD fields only when they are
"informative".  Unfortunately, even some uninformative MD fields are
essential, but we still don't want to generate them if we never had one
to begin with.

There's now an explicit representation of the fact that we read an MD
and parsed it correctly.  If so, we generate a new one, no matter how
informative it is.