redeye-single now generates BCF
First version that seems to generate readable and reasonably correct
BCF.  For the time being, it's uncompressed.  The order of likelihood
fields in the intermediate Avro containers had been changed to match the
order in VCF (which depends on the reference allele).

Indel records with multiple alleles are still broken.  Priors are not
applied, yet.