fixes bizarre memory leak in redeye-pileup
Was what leaked were continuations from the pileup module.  The leak was
(somehow) caused by 'eneeCheckIfDone' being used in 'pileup'.  This
commit fixes it by effectively inlining 'eneeCheckIfDone' into 'yield'
and simplifying it until no unused continuation was present anymore.
As a side effect, now 'yield' and 'pileup'' are in explicit CPS, but I'm
keeping them this way, both because the leak is plugged and because the
result doesn't look too imcomprehensible.