switches from JSON to binary
Parsing JSON was too expensive in terms of memory.  This patch replaces
the metadata format with one based on binary.  The price is that we
can't edit the file anymore (has to be done in GHCi) and that we cannot
easily extend it.  Maybe it should be upgraded to 'safecopy' or even
'acid-state'.